requirements.txt
python-telegram-bot==20.7
mutagen
ffmpeg-python
